Prototype/versioning - initial setup
