+++ prototype/src/app/assets/sass/application.scss
@@ -1,3 +1,5 @@
+@import 'v2.scss';
+
 //
 // For guidance on how to add CSS and SCSS see:
 // https://prototype-kit.service.gov.uk/docs/adding-css-javascript-and-images
