+++ prototype/src/app/assets/sass/v2.scss
@@ -1,0 +1,5 @@
+.v2 {
+    p {
+        color: red;
+    }
+}
