+++ prototype/src/app/views/latest/comparators/create/characteristics-post-mvp.html
