+++ prototype/src/app/views/latest/comparators/create/characteristics.html
