+++ prototype/src/app/views/latest/comparators/reset.html
