+++ prototype/src/app/views/latest/comparators/view-school.html
