+++ prototype/src/app/views/latest/comparators/view.html
