+++ prototype/src/app/views/latest/custom-data/school-characteristics.html
