+++ prototype/src/app/views/latest/templates/details-comparators.njk
