+++ prototype/src/app/views/latest/templates/priority-banner.njk
