+++ prototype/src/app/views/latest/templates/spend-category-details.njk
