+++ prototype/src/app/views/latest/templates/spending-panel.njk
