+++ prototype/src/app/views/layouts/latest.html
@@ -1,0 +1,4 @@
+{% extends "layouts/main.html" %}
+
+{% set bodyClasses = "latest" %}
+{% set version = "latest" %}
