+++ prototype/src/app/views/layouts/v1.html
@@ -1,0 +1,4 @@
+{% extends "layouts/main.html" %}
+
+{% set bodyClasses = "v1" %}
+{% set version = "v1" %}
