+++ prototype/src/app/views/layouts/v2.html
@@ -1,0 +1,4 @@
+{% extends "layouts/main.html" %}
+
+{% set bodyClasses = "v2" %}
+{% set version = "v2" %}
