+++ prototype/src/app/views/v1/index.html
@@ -1,0 +1,7 @@
+{% extends "layouts/v1.html" %}
+
+{% block content %}
+
+    <p>This is a page in version: {{ version }}</p>
+
+{% endblock %}
