+++ prototype/src/app/views/v2/index.html
@@ -1,0 +1,7 @@
+{% extends "layouts/v2.html" %}
+
+{% block content %}
+
+    <p>This is a page in version: {{ version }}</p>
+
+{% endblock %}
